fix(form/inputs): optimize rendering of member inputs
There were some issues here with member form inputs not being re-rendered properly when rendered inline in the editor
This makes sure we re-render those every time we need to (members change).

This can now be achieved by object equality checks after a0f93cba7430a4584d01256a55e078d52a05c459
